+++ eslint.config.mjs
@@ -1,0 +1,30 @@
+// @ts-check
+import globals from 'globals';
+import eslint from '@eslint/js';
+import tseslint from 'typescript-eslint';
+import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended';
+import eslintPlugin from 'eslint-plugin-eslint-plugin';
+import { defineConfig } from 'eslint/config';
+
+export default defineConfig(
+  eslint.configs.recommended,
+  ...tseslint.configs.recommended,
+  eslintPluginPrettierRecommended,
+  eslintPlugin.configs.recommended,
+  { ignores: ['node_modules', 'dist'] },
+  {
+    files: ['**/*.{ts,tsx}'],
+    languageOptions: { globals: globals.node, parserOptions: { project: true } },
+    rules: {
+      'no-undef': 'error',
+      'object-shorthand': 'warn',
+      'no-async-promise-executor': 'warn',
+      '@typescript-eslint/interface-name-prefix': 'off',
+      '@typescript-eslint/explicit-function-return-type': 'off',
+      '@typescript-eslint/explicit-module-boundary-types': 'off',
+      '@typescript-eslint/no-explicit-any': 'warn',
+      '@typescript-eslint/no-misused-promises': 'warn',
+      '@typescript-eslint/no-deprecated': 'error',
+    },
+  }
+);