+++ src/index.ts
@@ -1,7 +1,24 @@
+import { ESLint } from 'eslint';
 import { sortImportsRule } from './rules/sort-imports/sort-imports.rule';
 
-export = {
+const plugin = {
   rules: {
     import: sortImportsRule,
   },
-};
+} satisfies ESLint.Plugin;
+
+const configs = {
+  recommended: {
+    plugins: {
+      'beautiful-sort': plugin,
+    },
+    rules: {
+      'beautiful-sort/import': 'error',
+    },
+  },
+} satisfies ESLint.Plugin['configs'];
+
+export = {
+  ...plugin,
+  configs,
+} satisfies ESLint.Plugin;